Add /readyz endpoint
The ingress will occasionally attempt to hit new replicas, which aren't
ready yet. This results in an error message.

By adding a readinessProbe, the ingress won't hit the pod until it's
ready.

If we implement /readyz with the same cowboy router, the endpoints will
be ready at the same time.