Generate SSH host key during build
Previously, the build process relied on following the README file.
Without the host key, the application fails to start.

This change generates the host key during build.

You could argue that the host key should be generated by the Dockerfile,
or that it should be a K8s secret, or that it should be generated by an
init container.

This'll do for now.