Use K8s secret for SSH host key
If the image is made public, we probably *don't* want the host keys in
it.

If the image is deployed to multiple clusters, we probably want
different host keys on each one.

Right now, however, this is only a small step in the right direction:
because of the need for `kubectl port-forward`, all of the hosts are
`localhost`, so we don't do host key verification anyway.

It's still worth doing.