DOES NOT WORK
The CouchDB stuff allows for *this* node to connect to other nodes that
aren't using TLS. It doesn't allow for this node to *accept* some
connections without TLS.

To do that, we need to mess around with a custom implementation of
accept.

OR: there might be a way to run multiple distributions, and have
firewall rules forward in-pod connections to the non-TLS one.

OR: can we have epmd lie about the port for in-pod queries?

OR: can we (somehow) have nc in a sidecar that adds/removes TLS one way
or the other? Poor man's service mesh, if you like.